requirements.txt
absl-py==0.10.0
aniso8601==8.0.0
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.3.3
astunparse==1.6.3
async-generator==1.10
attrs==20.1.0
backcall==0.2.0
bleach==3.1.5
blis==0.4.1
bokeh==2.2.1
boto3==1.14.60
botocore==1.17.60
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
cvxpy==1.1.4
cycler==0.10.0
cymem==2.0.3
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
easydict==1.9
ecos==2.0.7.post1
entrypoints==0.3
erdos==0.3.0
-e git+https://github.com/erdos-project/pylot.git@382ca3d0b4e99eb01e777774287354f618e9676a#egg=erdos_pylot
fastcore==1.0.0
fastprogress==1.0.0
fastscript==1.0.0
filelock==3.0.12
filterpy==1.4.1
fire==0.3.1
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.0
future==0.18.2
gast==0.2.2
gdown==3.12.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.31.0
gunicorn==20.0.4
h5py==2.10.0
holoviews==1.13.3
idna==2.10
importlib-metadata==1.7.0
imutils==0.5.3
iniconfig==1.0.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-server==0.1.1
jupyterlab-pygments==0.1.1
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lapsolver==1.1.0
lazy-object-proxy==1.4.3
llvmlite==0.33.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==2.2.4
mccabe==0.6.1
metaflow==2.2.3
mistune==0.8.4
more-itertools==8.4.0
murmurhash==1.0.2
nbconvert==5.6.1
nbdev==1.0.1
nbformat==5.0.7
networkx==2.2
notebook==6.1.3
numba==0.50.0
numpy==1.16.6
oauthlib==3.1.0
open3d-python==0.5.0.0
opencv-contrib-python==4.4.0.42
opencv-python==4.4.0.42
opt-einsum==3.3.0
osqp==0.6.1
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
panel==0.9.7
param==1.9.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
plotly==4.11.0
pluggy==0.13.1
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.13.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyct==0.4.8
pygame==1.9.6
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
pytesseract==0.3.6
pytest==6.0.1
python-dateutil==2.8.1
pytz==2020.1
pyviz-comms==0.7.6
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
s3transfer==0.3.3
scikit-image==0.14.5
scikit-learn==0.20.0
scipy==1.4.1
scs==2.1.2
seaborn==0.11.0
Send2Trash==1.5.0
Shapely==1.6.4
six==1.15.0
spacy==2.3.2
SQLAlchemy==1.3.19
srsly==1.0.2
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==7.4.1
threadpoolctl==2.1.0
toml==0.10.1
torch==1.3.1
torchvision==0.2.1
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
voila==0.1.23
wasabi==0.7.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
